// src/types/types.ts
export interface QuizSetLeaderBoardSummaryPayload {
  uid: string;
  category_name: string;
  question_players: QsetPlayer[];
}

interface QsetPlayer {
  id: number;
  q_player: QPlayer;
  points: number;
  timer: number;
  score?: number;
}

interface QPlayer {
  id: number;
  first_name: string;
  last_name: string;
  profile_pic?: string | null;
}